rt: add experimental ray model cache eviction
reasoning is that we might have too many different objects/model in the scene and run out of cache slots early, even though the number of currently active/drawn objects is low.

but with increased cache size (which was needed for high currently drawn object count) the eviction code wasn't even reached.