vk: fix unintended clearing of images on barriers
Previously we forced src image layout to be UNDEFINED if the image was
to be written into. This lead to RADV driver to completely clear our so
painfully constructed ray traced frame.

The correct layout transition should probably be something like this: if
we're not to _read_ from image contents, only then we can be sure that
its contents are not needed anymore, and can be discarded by settind the
src layout to UNDEFINED.