vk: start adding new global vars for refapi
This is, as many surrounding commits are, is not a self-sufficient
commit. These commits are here just to logically roughly separate a
bunch of big changes. Each commit doesn't compile by itself.

This just introduces the new global variables, but doesn't update all
their users. This will come in a series of further commits.